requirements.txt
Cython==0.28.5
Flask==1.0.2
graphviz==0.8.4
h5py==2.8.0
#hdbscan==0.8.15
html5lib==0.9999999
httplib2==0.9.1
Keras==2.1.4
matplotlib==2.1.2
numpy==1.14.1
opencv-python==3.4.2.17
pandas==0.22.0
Pillow==5.2.0
requests==2.9.1
scikit-image==0.14.0
scikit-learn==0.19.1
scipy==1.0.0
tensorboard==1.6.0
tensorflow==1.6.0
urllib3==1.13.1
utils==0.9.0
